create classes for wards
